Calendar in 2 components and swipe test!
