+++ src/app/calendar/calendar.component.html
@@ -1,54 +1,4 @@
 <h1>Calendar</h1>
 <div class="calendar">
-  <div class="calendar-nav">
-    <!--button>Month</button-->
-    <div>Show schedule by week</div>
-    <button (click)="this.day('7777777')" onclick="location.reload();">Week</button>
-    <button (click)="this.day('333')" onclick="location.reload();">3 Days</button>
-    <button (click)="this.day('1')" onclick="location.reload();">Day</button>
-  </div>
-  <div class="calendar-day" *ngFor="let day of view">
-    <hr class="timeline" style="position: ;">
-    <div class="calendar-header">
-      <div class="calendar-weekday">{{ day.date | date:'EEEE dd.MM.y' }}</div>
-    </div>
-
-    <div class="lecture-body" *ngFor="let lecture of day.lectures">
-      <!-- calendar lectures -->
-      <div class="lecture-time">{{lecture.time.start}} - {{lecture.time.end}}</div>
-      <div class="lecture-information" *ngIf="lecture.check==''" (click)="lecture.onCheck()">
-        <!-- div jakaa tiedon hyvin eri riveille ja sillä voi myös luokitella css tyylejä tarvittaessa -->
-        <div>{{lecture.course.id}}</div>
-        <div>{{lecture.course.name}}</div>
-        <div>{{lecture.course.room}}</div>
-      </div>
-      <div class="lecture-information lecture-check-true" *ngIf="lecture.check=='true'" (click)="lecture.onCheck()">
-        <!-- div jakaa tiedon hyvin eri riveille ja sillä voi myös luokitella css tyylejä tarvittaessa -->
-        <div>{{lecture.course.id}}</div>
-        <div>{{lecture.course.name}}</div>
-        <div>{{lecture.course.room}}</div>
-      </div>
-      <div class="lecture-information lecture-check-false" *ngIf="lecture.check=='false'" (click)="lecture.onCheck()">
-        <!-- div jakaa tiedon hyvin eri riveille ja sillä voi myös luokitella css tyylejä tarvittaessa -->
-        <div>{{lecture.course.id}}</div>
-        <div>{{lecture.course.name}}</div>
-        <div>{{lecture.course.room}}</div>
-      </div>
-    </div>
-
-    <div *ngFor="let lecture of day.lectures">
-        <md-card class="lecture-check-true">
-            <md-card-header>
-                <md-card-title>{{lecture.time.start}} - {{lecture.time.end}}</md-card-title>
-                <!--md-card-subtitle><hr></md-card-subtitle-->
-            </md-card-header>
-            <md-card-content>
-                <div>{{lecture.course.id}}</div>
-                <div>{{lecture.course.name}}</div>
-                <div>{{lecture.course.room}}</div>
-            </md-card-content>
-            <!--md-card-actions></md-card-actions-->
-        </md-card>
-    </div>
-  </div>
+  <calendarview></calendarview>
 </div>
