+++ src/app/template/calendarview/calendarview.component.css
@@ -1,0 +1,83 @@
+.calendar::selection {
+color: red;
+background: white;
+}
+
+.calendar::-moz-selection {
+color: red;
+background: white;
+}
+
+.calendar {
+
+}
+
+.calendar-header {
+  border-style: solid;
+  border-color: black;
+  border-width: 0 0 1px 1px;
+  margin: 5px;
+}
+
+.calendar-day {
+  /* inline-block
+  Täyttää vain tarvittavan verran ja tavaraa voi olla vierekkäin riveissä.
+
+  */
+  display: inline-block;
+  padding: 5px;
+}
+
+.calendar-weekday {
+  padding: 3px;
+  font-weight: bold;
+}
+
+.lecture-body {
+  /*Vastaa <- ja _ borderista*/
+  border-top-left-radius: 8px;
+  border-width: 0;
+  border-style: solid;
+  border-color: black;
+  border-left-width: 1px;
+  border-bottom-width: 1px;
+  margin: 5px;
+}
+
+.lecture-time {
+  /*Vastaa omasta ^ ja -> borderista*/
+  display: inline; /*piirtää suoraan riville*/
+  border-top-right-radius: 8px;
+  border-top-left-radius: 8px;
+  padding: 3px;
+  border-top: 1px solid black;
+  border-right: 1px solid black;
+}
+
+.lecture-information {
+  /*
+  Vastaa ^, -> borderista
+
+  */
+  padding: 3px;
+  text-align: center;
+  border-top: 1px solid black;
+  border-right: 1px solid black;
+}
+
+md-card {
+    background-color: lightgray;
+}
+.lecture-check-true {
+  /*
+  Vastaa myös laidan border väristä
+  */
+  border-left: 15px solid #5cb85c;
+}
+
+.lecture-check-false {
+  /*
+  Vastaa myös laidan border väristä
+  */
+  border-left: 15px solid #d9534f;
+}
