+++ src/app/template/calendarview/index.ts
@@ -1,0 +1,1 @@
+export * from './calendarview.component';
